Исправление картинок

+++ src/main/java/ru/skypro/homework/dto/AdsDto.java
@@ -8,7 +8,7 @@ public class AdsDto {
 
     private Integer author;
     private String image;
-    private Integer pk;
+    private Long pk;
     private Integer price;
     private String title;
